@@ -12,69 +12,76 @@
     "ionic:serve": "ionic-app-scripts serve"
   },
   "dependencies": {
-    "@angular/common": "4.1.3",
-    "@angular/compiler": "4.1.3",
-    "@angular/compiler-cli": "4.1.3",
-    "@angular/core": "4.1.3",
-    "@angular/forms": "4.1.3",
-    "@angular/http": "4.1.3",
-    "@angular/platform-browser": "4.1.3",
-    "@angular/platform-browser-dynamic": "4.1.3",
-    "@ionic-native/barcode-scanner": "^4.1.0",
-    "@ionic-native/core": "3.12.1",
-    "@ionic-native/facebook": "^4.2.1",
-    "@ionic-native/in-app-browser": "^4.2.1",
-    "@ionic-native/splash-screen": "3.12.1",
-    "@ionic-native/status-bar": "3.12.1",
-    "@ionic/cloud-angular": "^0.12.0",
-    "@ionic/storage": "2.0.1",
-    "DniScanner": "file:///Users/alexmarcos/Documents/Projects/VIDChain/DniScanner",
-    "cordova-android": "^6.2.3",
-    "cordova-plugin-compat": "^1.1.0",
-    "cordova-plugin-console": "^1.0.5",
-    "cordova-plugin-device": "1.1.4",
-    "cordova-plugin-dni-scanner-nfc": "/Users/alexmarcos/Documents/Projects/VIDChain/DniScanner",
+    "@angular/common": "5.0.3",
+    "@angular/compiler": "5.0.3",
+    "@angular/compiler-cli": "5.0.3",
+    "@angular/core": "5.0.3",
+    "@angular/forms": "5.0.3",
+    "@angular/http": "5.0.3",
+    "@angular/platform-browser": "5.0.3",
+    "@angular/platform-browser-dynamic": "5.0.3",
+    "@ionic-native/barcode-scanner": "^4.5.3",
+    "@ionic-native/core": "4.4.0",
+    "@ionic-native/facebook": "^4.5.3",
+    "@ionic-native/firebase": "^4.5.3",
+    "@ionic-native/google-plus": "^4.5.3",
+    "@ionic-native/in-app-browser": "^4.5.3",
+    "@ionic-native/splash-screen": "4.4.0",
+    "@ionic-native/status-bar": "4.4.0",
+    "@ionic/pro": "^1.0.20",
+    "@ionic/storage": "2.1.3",
+    "cordova-android": "6.4.0",
+    "cordova-plugin-browsertab": "^0.2.0",
+    "cordova-plugin-buildinfo": "^2.0.1",
+    "cordova-plugin-compat": "^1.2.0",
+    "cordova-plugin-device": "^2.0.1",
+    "cordova-plugin-dni-scanner-nfc": "./plugins/cordova-plugin-dni-scanner-nfc",
     "cordova-plugin-facebook4": "^1.7.4",
-    "cordova-plugin-inappbrowser": "^1.7.1",
-    "cordova-plugin-splashscreen": "~4.0.1",
-    "cordova-plugin-statusbar": "2.2.2",
-    "cordova-plugin-template": "file:///Users/alexmarcos/Downloads/dniScanEmpty",
-    "cordova-plugin-whitelist": "1.3.1",
-    "ionic-angular": "3.6.0",
-    "ionic-plugin-keyboard": "~2.2.1",
+    "cordova-plugin-firebase": "^0.1.25",
+    "cordova-plugin-googleplus": "^5.2.1",
+    "cordova-plugin-inappbrowser": "^2.0.2",
+    "cordova-plugin-ionic-keyboard": "^2.0.5",
+    "cordova-plugin-ionic-webview": "^1.1.16",
+    "cordova-plugin-push": "0.0.1",
+    "cordova-plugin-splashscreen": "^5.0.2",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "cordova-universal-links-plugin": "^1.2.1",
+    "firebase": "^4.10.1",
+    "ionic-angular": "3.9.2",
     "ionicons": "3.0.0",
-    "moment": "^2.18.1",
-    "phonegap-plugin-barcodescanner": "^6.0.8",
-    "phonegap-plugin-push": "^2.0.0",
-    "rxjs": "5.4.0",
-    "sha.js": "^2.4.9",
+    "moment": "^2.20.1",
+    "phonegap-plugin-barcodescanner": "^7.0.2",
+    "rxjs": "5.5.2",
+    "sha.js": "^2.4.10",
     "sw-toolbox": "3.6.0",
-    "zone.js": "0.8.12"
+    "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "2.1.3",
-    "ionic": "3.9.2",
-    "typescript": "2.3.4"
+    "@ionic/app-scripts": "^3.1.8",
+    "typescript": "2.4.2"
   },
   "description": "An Ionic project",
   "cordova": {
     "plugins": {
-      "cordova-plugin-console": {},
+      "cordova-plugin-inappbrowser": {},
+      "cordova-plugin-whitelist": {},
       "cordova-plugin-device": {},
       "cordova-plugin-splashscreen": {},
-      "cordova-plugin-statusbar": {},
-      "cordova-plugin-whitelist": {},
-      "ionic-plugin-keyboard": {},
+      "cordova-plugin-ionic-webview": {},
+      "cordova-plugin-ionic-keyboard": {},
+      "cordova-plugin-dni-scanner-nfc": {},
+      "cordova-plugin-buildinfo": {},
+      "cordova-universal-links-plugin": {},
+      "cordova-plugin-browsertab": {},
       "phonegap-plugin-barcodescanner": {},
-      "cordova-plugin-inappbrowser": {},
       "cordova-plugin-facebook4": {
-        "APP_ID": "2011955215702503",
-        "APP_NAME": "vidchain"
+        "APP_ID": "1600429613357545",
+        "APP_NAME": "ViDChainAPP"
       },
-      "phonegap-plugin-push": {
-        "SENDER_ID": "636222461475"
+      "cordova-plugin-googleplus": {
+        "REVERSED_CLIENT_ID": "938949312980-f79pafv92qqtr45i7f72orhv8ig0kq3v.apps.googleusercontent.com"
       },
-      "cordova-plugin-dni-scanner-nfc": {}
+      "cordova-plugin-firebase": {}
     },
     "platforms": [
       "android"
