@@ -1,4 +1,3 @@
-import {Http} from '@angular/http';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/toPromise';
 import { Component } from '@angular/core';
@@ -6,6 +5,7 @@ import { BarcodeScanner } from '@ionic-native/barcode-scanner';
 import CONSTANT from "../../constants";
 import {SendAttributesPage} from "../sendAttributes/sendAttributes";
 import {ModalController} from "ionic-angular";
+import { HttpClient } from '@angular/common/http';
 @Component({
   selector: 'page-login',
   templateUrl: 'login.html'
@@ -15,7 +15,7 @@ export class LoginPage {
   private URL_CONFIRM_LOGIN = CONSTANT.URL.URL_CONFIRM_LOGIN;
   constructor(private barcodeScanner: BarcodeScanner,
               public modalCtrl: ModalController,
-              private http:Http) {
+              private http:HttpClient) {
 
   }
 
@@ -24,11 +24,11 @@ export class LoginPage {
       showTorchButton : true,
       showFlipCameraButton : true
     };
-
+    
     this.barcodeScanner.scan(options).then((barcodeData) => {
       if(!barcodeData.cancelled && barcodeData.text.indexOf('vidchain') != -1){
         this.URL_CONFIRM_LOGIN = CONSTANT.URL.URL_CONFIRM_LOGIN;
-        this.http.get(this.URL_CONFIRM_LOGIN+'?id='+barcodeData.text).map(res => res.json()).subscribe(data => {
+        this.http.get(this.URL_CONFIRM_LOGIN+'?id='+barcodeData.text).subscribe(data => {
           let sendAttributesModal = this.modalCtrl.create(SendAttributesPage,{attrToSend:data});
           sendAttributesModal.present();
         });
@@ -36,6 +36,7 @@ export class LoginPage {
     }, (err) => {
       // An error occurred
     });
+    
   }
 
 }
